refactor: 迁移 FeishuRenderer 逻辑到 FeishuPlugin
将 FeishuRenderer 的卡片构建逻辑（build_card、build_text、
dispatch_blocks、should_use_card 等）迁移到 FeishuPlugin。
FeishuPlugin 覆盖 IMPlugin::render() 实现飞书平台差异化渲染，
包括 Thinking/ToolUse/ToolResult 等非文本块的卡片元素构建。
删除独立的 renderer/feishu.rs 及其测试文件。

Source: user
Type: refactor
